requirements.txt
Flask~=2.0
python-dotenv~=0.13
psycopg2-binary~=2.9
requests~=2.19

